requirements.txt
djangorestframework-jwt
python-weixin
certifi==2018.4.16
Django==2.0.5
django-model-utils==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.4
